src/backend/requirements.txt

flask==2.3.2
flask-cors==3.0.10
docx2txt==0.8
spacy==3.5.3
pdfminer.six==20221105
